Finished implementation of read()
The `read()' function reads the file given in the path and fills the `problem' struct accordingly. The file encodes the set cover problem and is formatted like this:

<number of elements> <number of sets (M)>
<cost of set 1> <element 0 of set 1> ... <element n_1 of set 1>
...
<cost of set M> <element 0 of set M> ... <element n_M of set M>

Delimiting character is blank.
File may end with line break or without.
Successive blanks will be treated as one.

Tested with `sc_6_1' and `sc_25_0'.