Added remove element functionality
`remove_element_from_all_sets.c' removes an element from all sets given in the element-value-table for the element's value.
E.g. if the element has a value 2, the index-2 entry of the element-value-table is read which returns an array of (set index, element index) pairs.
Every pair is used to address an element in a set.
The previous and next element will be newly linked, s.t. the element is effectively removed.
When iterating through the elements of this set, starting at `index_of_root_element' of a `set', this element cannot be reached with the consecutive use of `index_next' indices.
Special cases of first element, last element, and only remaining element are taken care of.