Implemented greedy with look-up table solver
`greedy_with_LUT()' takes a path as input argument, and returns the solution string.
The solution string looks something like this for problem `sc_6_1':

3 0
0 1 0 1 1 0

The first line has two integers.
The first integer is the overall cost.
The second integer is 1 if the optimum is known, 0 otherwise.
The second line is the bit mask of selected sets.

Complies with the handout "Discrete Optimization: Set Cover" for the optional programming assignment in week 3.